% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{bsf_car}
\alias{bsf_car}
\title{Boostrap SMC for CAR model with Binomial observations}
\usage{
bsf_car(nnbs, nbs, tau, d, y, u, idx, mu = 0, use_mu = TRUE,
  n_particles, ess_threshold = 2, reorder = TRUE,
  seed = sample(.Machine$integer.max, size = 1))
}
\arguments{
\item{nnbs}{Vector defining the number of neighbours for each vertex.}

\item{nbs}{Matrix of indices of defining neighbours for each vertex.}

\item{tau}{Precision parameter for CAR field.}

\item{d}{Properness parameter for the CAR field.}

\item{y}{Vector of observations.}

\item{u}{Vector of trials. Default to 1.}

\item{idx}{Vector defining the dependencies between y and x.}

\item{mu}{Intercept of the linear predictor. Defaults to 0.}

\item{use_mu}{Use mu in the model or not. Default is \code{TRUE}.}

\item{n_particles}{Number of particles for SMC.}

\item{ess_threshold}{Resampling is done when the effective sample size estimator is less than this threshold times the number of particles.
For example if \code{ess_threshold=1} resampling is done at each iteration, whereas if \code{ess_threshold=0} the algorithm reduces to simple importance sampling.}

\item{reorder}{If \code{TRUE} (default), reordering is performed for increased efficiency.}

\item{seed}{Seed for the random number generator.}
}
\description{
Boostrap SMC for CAR model with Binomial observations
}
