% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_binomial_car}
\alias{mcmc_binomial_car}
\title{MCMC for Binomial CAR model}
\usage{
mcmc_binomial_car(nnbs, nbs, tau, d, y, u, idx, mu = 0, use_mu = TRUE,
  n_iter, n_burnin, n_particles = 0, initial_mode, max_iter = 100,
  conv_tol = 1e-08, seed = sample(.Machine$integer.max, size = 1),
  S = NULL, ratio_correction = TRUE, delayed_acceptance = TRUE,
  reorder = TRUE, ess_threshold = 1)
}
\arguments{
\item{nnbs}{Vector defining the number of neighbours for each vertex.}

\item{nbs}{Matrix of indices of defining neighbours for each vertex.}

\item{y}{Vector of observations.}

\item{u}{Vector of exposures. Default to 1.}

\item{idx}{Vector defining the dependencies between y and x.}

\item{mu}{Intercept of the linear predictor. Defaults to 1.}

\item{n_particles}{Number of particles for SMC.}

\item{initial_mode}{Initial mode estimate of x.}

\item{max_iter}{Maximum number of iterations for the approximation algorithm.}

\item{conv_tol}{Tolerance parameter for the approximation algorithm.}

\item{phi}{Autoregressive parameter of the latent process.}

\item{sigma}{Standard deviation of the AR process noise.}

\item{nu}{Intercept term of the AR process (stationary mean is nu / (1 - phi)).}
}
\description{
MCMC for Binomial CAR model
}
