% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefield-package.R
\docType{package}
\name{particlefield}
\alias{particlefield}
\alias{particlefield-package}
\title{Sequential Monte Carlo Methods for Markov Random Fields}
\description{
This package contains codes for running sequential Monte Carlo (SMC) for 
simple latent Gaussian Markov random field (GMRF) where the observations are 
Poisson or binomial. The methodology is introduced in Lindsten, Helske, Vihola (2018).
}
