% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_graph.R
\name{print_graph}
\alias{print_graph}
\title{Print the graph structure to console}
\usage{
print_graph(nnbs, nbs)
}
\arguments{
\item{nnbs}{Vector defining the number of neighbours for each vertex.}

\item{nbs}{Matrix of indices of defining neighbours for each vertex.}
}
\description{
Print the graph structure to console
}
\examples{
# Graph on line with 10 vertices
nnbs <- c(1, rep(2, 8), 1)
nbs <- matrix(0, 10, 10)
nbs[1, 1] <- 2
for (i in 2:9) nbs[i, 1:2] <- c(i - 1, i + 1)
nbs[10, 1] <- 9
print_graph(nnbs, nbs)

}
