% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_apf.R
\name{psi_car}
\alias{psi_car}
\title{psi-APF for Poisson model with latent AR(1) process}
\usage{
psi_car(nnbs, nbs, tau, d, y, u, idx, mu = 0, use_mu = TRUE,
  n_particles, initial_mode, max_iter = 100, conv_tol = 1e-08,
  seed = sample(.Machine$integer.max, size = 1), reorder = TRUE,
  ess_threshold = NULL)
}
\arguments{
\item{nnbs}{Vector defining the number of neighbours for each vertex.}

\item{nbs}{Matrix of indices of defining neighbours for each vertex.}

\item{y}{Vector of observations.}

\item{u}{Vector of exposures. Default to 1.}

\item{idx}{Vector defining the dependencies between y and x.}

\item{mu}{Intercept of the linear predictor. Defaults to 1.}

\item{n_particles}{Number of particles for SMC.}

\item{initial_mode}{Initial mode estimate of x.}

\item{max_iter}{Maximum number of iterations for the approximation algorithm.}

\item{conv_tol}{Tolerance parameter for the approximation algorithm.}

\item{phi}{Autoregressive parameter of the latent process.}

\item{sigma}{Standard deviation of the AR process noise.}

\item{nu}{Intercept term of the AR process (stationary mean is nu / (1 - phi)).}
}
\description{
psi-APF for Poisson model with latent AR(1) process
}
